chỉnh sửa background và một chút hiệu ứng
bản này chưa chính thức nha tại tui bị lỗi chút đang sửa dần
