+++ frontend/src/pages/auth/Login/Index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Login";
