--- frontend/src/pages/auth/Register.tsx
@@ -1,367 +1,0 @@
-import { useState } from "react";
-import { Link, useNavigate } from "react-router-dom";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { Checkbox } from "@/components/ui/checkbox";
-import { Separator } from "@/components/ui/separator";
-import {
-  Car,
-  Eye,
-  EyeOff,
-  Mail,
-  Lock,
-  User,
-  Phone,
-  Calendar,
-} from "lucide-react";
-import { useToast } from "@/hooks/use-toast";
-import { useTranslation } from "@/contexts/TranslationContext";
-
-interface User {
-  id: string;
-  name: string;
-  email: string;
-  role: "customer" | "staff" | "admin";
-  stationId?: string;
-}
-
-interface RegisterProps {
-  onRegister: (userData: User) => void;
-}
-
-const Register = ({ onRegister }: RegisterProps) => {
-  const { t } = useTranslation();
-  const [formData, setFormData] = useState({
-    fullName: "",
-    email: "",
-    phone: "",
-    dateOfBirth: "",
-    password: "",
-    confirmPassword: "",
-  });
-  const [showPassword, setShowPassword] = useState(false);
-  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
-  const [agreeToTerms, setAgreeToTerms] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-  const navigate = useNavigate();
-  const { toast } = useToast();
-
-  const handleInputChange = (field: string, value: string) => {
-    setFormData((prev) => ({
-      ...prev,
-      [field]: value,
-    }));
-  };
-
-  const handleSubmit = async (e: React.FormEvent) => {
-    e.preventDefault();
-
-    if (!agreeToTerms) {
-      toast({
-        title: t("register.agreementRequired"),
-        description: t("register.agreeToTerms"),
-        variant: "destructive",
-      });
-      return;
-    }
-
-    if (formData.password !== formData.confirmPassword) {
-      toast({
-        title: t("register.passwordMismatch"),
-        description: t("register.passwordsDoNotMatch"),
-        variant: "destructive",
-      });
-      return;
-    }
-
-    setIsLoading(true);
-
-    // Mock registration - in real app, this would call an API
-    setTimeout(() => {
-      if (formData.fullName && formData.email && formData.password) {
-        const userData: User = {
-          id: `user_${Date.now()}`,
-          name: formData.fullName,
-          email: formData.email,
-          role: "customer", // Default role for new registrations
-        };
-        onRegister(userData);
-        toast({
-          title: t("register.welcome"),
-          description: t("register.accountCreated"),
-        });
-        navigate("/dashboard");
-      } else {
-        toast({
-          title: t("register.error"),
-          description: t("register.fillAllFields"),
-          variant: "destructive",
-        });
-      }
-      setIsLoading(false);
-    }, 1000);
-  };
-
-  return (
-    <div className="min-h-screen bg-gradient-hero flex items-center justify-center p-4">
-      <div className="w-full max-w-md">
-        {/* Logo */}
-        <div className="text-center mb-8">
-          <Link to="/" className="inline-flex items-center space-x-2 group">
-            <div className="p-3 bg-white rounded-xl group-hover:scale-105 transition-transform duration-200 shadow-lg">
-              <Car className="h-8 w-8 text-primary" />
-            </div>
-            <span className="text-2xl font-bold text-white">EVRentals</span>
-          </Link>
-        </div>
-
-        {/* Register Card */}
-        <Card className="shadow-premium">
-          <CardHeader className="space-y-1">
-            <CardTitle className="text-2xl font-bold text-center">
-              {t("register.createAccount")}
-            </CardTitle>
-            <CardDescription className="text-center">
-              {t("register.joinElectric")}
-            </CardDescription>
-          </CardHeader>
-          <CardContent className="space-y-6">
-            <form onSubmit={handleSubmit} className="space-y-4">
-              {/* Full Name */}
-              <div className="space-y-2">
-                <Label htmlFor="fullName">{t("register.fullName")} *</Label>
-                <div className="relative">
-                  <User className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
-                  <Input
-                    id="fullName"
-                    type="text"
-                    placeholder={t("register.fullNamePlaceholder")}
-                    value={formData.fullName}
-                    onChange={(e) =>
-                      handleInputChange("fullName", e.target.value)
-                    }
-                    className="pl-10 text-black"
-                    required
-                  />
-                </div>
-              </div>
-
-              {/* Email */}
-              <div className="space-y-2">
-                <Label htmlFor="email">{t("register.emailAddress")} *</Label>
-                <div className="relative">
-                  <Mail className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
-                  <Input
-                    id="email"
-                    type="email"
-                    placeholder={t("register.emailPlaceholder")}
-                    value={formData.email}
-                    onChange={(e) => handleInputChange("email", e.target.value)}
-                    className="pl-10 text-black"
-                    required
-                  />
-                </div>
-              </div>
-
-              {/* Phone */}
-              <div className="space-y-2">
-                <Label htmlFor="phone">{t("register.phoneNumber")} *</Label>
-                <div className="relative">
-                  <Phone className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
-                  <Input
-                    id="phone"
-                    type="tel"
-                    placeholder={t("register.phonePlaceholder")}
-                    value={formData.phone}
-                    onChange={(e) => handleInputChange("phone", e.target.value)}
-                    className="pl-10 text-black"
-                    required
-                  />
-                </div>
-              </div>
-
-              {/* Date of Birth */}
-              <div className="space-y-2">
-                <Label htmlFor="dateOfBirth">
-                  {t("register.dateOfBirth")} *
-                </Label>
-                <div className="relative">
-                  <Calendar className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
-                  <Input
-                    id="dateOfBirth"
-                    type="date"
-                    value={formData.dateOfBirth}
-                    onChange={(e) =>
-                      handleInputChange("dateOfBirth", e.target.value)
-                    }
-                    className="pl-10 text-black"
-                    required
-                  />
-                </div>
-              </div>
-
-              {/* Password */}
-              <div className="space-y-2">
-                <Label htmlFor="password">{t("register.password")} *</Label>
-                <div className="relative">
-                  <Lock className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
-                  <Input
-                    id="password"
-                    type={showPassword ? "text" : "password"}
-                    placeholder={t("register.passwordPlaceholder")}
-                    value={formData.password}
-                    onChange={(e) =>
-                      handleInputChange("password", e.target.value)
-                    }
-                    className="pl-10 pr-10 text-black"
-                    required
-                  />
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    size="sm"
-                    className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
-                    onClick={() => setShowPassword(!showPassword)}
-                  >
-                    {showPassword ? (
-                      <EyeOff className="h-4 w-4 text-muted-foreground" />
-                    ) : (
-                      <Eye className="h-4 w-4 text-muted-foreground" />
-                    )}
-                  </Button>
-                </div>
-              </div>
-
-              {/* Confirm Password */}
-              <div className="space-y-2">
-                <Label htmlFor="confirmPassword">
-                  {t("register.confirmPassword")} *
-                </Label>
-                <div className="relative">
-                  <Lock className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
-                  <Input
-                    id="confirmPassword"
-                    type={showConfirmPassword ? "text" : "password"}
-                    placeholder={t("register.confirmPasswordPlaceholder")}
-                    value={formData.confirmPassword}
-                    onChange={(e) =>
-                      handleInputChange("confirmPassword", e.target.value)
-                    }
-                    className="pl-10 pr-10 text-black"
-                    required
-                  />
-                  <Button
-                    type="button"
-                    variant="ghost"
-                    size="sm"
-                    className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
-                    onClick={() => setShowConfirmPassword(!showConfirmPassword)}
-                  >
-                    {showConfirmPassword ? (
-                      <EyeOff className="h-4 w-4 text-muted-foreground" />
-                    ) : (
-                      <Eye className="h-4 w-4 text-muted-foreground" />
-                    )}
-                  </Button>
-                </div>
-              </div>
-
-              {/* Terms Agreement */}
-              <div className="flex items-center space-x-2">
-                <Checkbox
-                  id="terms"
-                  checked={agreeToTerms}
-                  onCheckedChange={(checked) =>
-                    setAgreeToTerms(checked as boolean)
-                  }
-                />
-                <label htmlFor="terms" className="text-sm leading-relaxed">
-                  {t("register.agreeToTermsText")}{" "}
-                  <Link
-                    to="/terms"
-                    className="text-primary hover:text-primary-dark"
-                  >
-                    {t("register.termsOfService")}
-                  </Link>{" "}
-                  {t("register.and")}{" "}
-                  <Link
-                    to="/privacy"
-                    className="text-primary hover:text-primary-dark"
-                  >
-                    {t("register.privacyPolicy")}
-                  </Link>
-                </label>
-              </div>
-
-              {/* Submit Button */}
-              <Button
-                type="submit"
-                className="w-full btn-hero"
-                disabled={isLoading || !agreeToTerms}
-              >
-                {isLoading
-                  ? t("common.creatingAccount")
-                  : t("common.createAccount")}
-              </Button>
-            </form>
-
-            <Separator />
-
-            {/* Social Registration (Mock) */}
-            <div className="space-y-3">
-              <Button variant="outline" className="w-full" disabled>
-                <svg className="h-5 w-5 mr-2" viewBox="0 0 24 24">
-                  <path
-                    fill="currentColor"
-                    d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z"
-                  />
-                  <path
-                    fill="currentColor"
-                    d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z"
-                  />
-                  <path
-                    fill="currentColor"
-                    d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z"
-                  />
-                  <path
-                    fill="currentColor"
-                    d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z"
-                  />
-                </svg>
-                Sign up with Google
-              </Button>
-            </div>
-
-            {/* Sign In Link */}
-            <div className="text-center text-sm">
-              <span className="text-muted-foreground">
-                {t("register.alreadyHaveAccount")}{" "}
-              </span>
-              <Link
-                to="/login"
-                className="text-primary hover:text-primary-dark font-medium transition-colors"
-              >
-                {t("register.signInHere")}
-              </Link>
-            </div>
-          </CardContent>
-        </Card>
-
-        {/* Demo Note */}
-        <div className="mt-6 text-center">
-          <p className="text-white/80 text-sm">{t("register.demoNote")}</p>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default Register;
