+++ frontend/src/pages/auth/Register/RegisterForm.tsx
@@ -1,0 +1,252 @@
+import { useState } from "react";
+import { Link, useNavigate } from "react-router-dom";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Button } from "@/components/ui/button";
+import { Card, CardContent, CardHeader, CardTitle, CardDescription } from "@/components/ui/card";
+import { Checkbox } from "@/components/ui/checkbox";
+import { Separator } from "@/components/ui/separator";
+import { Car, Eye, EyeOff, Mail, Lock, User, Phone, Calendar } from "lucide-react";
+import { useToast } from "@/hooks/use-toast";
+import { useTranslation } from "@/contexts/TranslationContext";
+import { User as UserType } from "./types";
+// Stylesheet removed (file not present) to avoid build error
+
+interface RegisterFormProps {
+  onRegister: (userData: UserType) => void;
+}
+
+export const RegisterForm = ({ onRegister }: RegisterFormProps) => {
+  const { t } = useTranslation();
+  const { toast } = useToast();
+  const navigate = useNavigate();
+
+  const [formData, setFormData] = useState({
+    fullName: "",
+    email: "",
+    phone: "",
+    dateOfBirth: "",
+    password: "",
+    confirmPassword: "",
+  });
+
+  const [showPassword, setShowPassword] = useState(false);
+  const [showConfirmPassword, setShowConfirmPassword] = useState(false);
+  const [agreeToTerms, setAgreeToTerms] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
+
+  const handleInputChange = (field: string, value: string) => {
+    setFormData((prev) => ({ ...prev, [field]: value }));
+  };
+
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault();
+
+    if (!agreeToTerms) {
+      toast({
+        title: t("register.agreementRequired"),
+        description: t("register.agreeToTerms"),
+        variant: "destructive",
+      });
+      return;
+    }
+
+    if (formData.password !== formData.confirmPassword) {
+      toast({
+        title: t("register.passwordMismatch"),
+        description: t("register.passwordsDoNotMatch"),
+        variant: "destructive",
+      });
+      return;
+    }
+
+    setIsLoading(true);
+
+    setTimeout(() => {
+      if (formData.fullName && formData.email && formData.password) {
+        const userData: UserType = {
+          id: `user_${Date.now()}`,
+          name: formData.fullName,
+          email: formData.email,
+          role: "customer",
+        };
+        onRegister(userData);
+        toast({
+          title: t("register.welcome"),
+          description: t("register.accountCreated"),
+        });
+        navigate("/dashboard");
+      } else {
+        toast({
+          title: t("register.error"),
+          description: t("register.fillAllFields"),
+          variant: "destructive",
+        });
+      }
+      setIsLoading(false);
+    }, 1000);
+  };
+
+  return (
+    <Card className="shadow-premium w-full max-w-md">
+      <CardHeader className="space-y-1">
+        <CardTitle className="text-2xl font-bold text-center">
+          {t("register.createAccount")}
+        </CardTitle>
+        <CardDescription className="text-center">
+          {t("register.joinElectric")}
+        </CardDescription>
+      </CardHeader>
+      <CardContent className="space-y-6">
+        <form onSubmit={handleSubmit} className="space-y-4">
+          {/* Full Name */}
+          <div className="space-y-2">
+            <Label htmlFor="fullName">{t("register.fullName")} *</Label>
+            <div className="relative">
+              <User className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
+              <Input
+                id="fullName"
+                type="text"
+                placeholder={t("register.fullNamePlaceholder")}
+                value={formData.fullName}
+                onChange={(e) => handleInputChange("fullName", e.target.value)}
+                className="pl-10 text-black"
+                required
+              />
+            </div>
+          </div>
+
+          {/* Email */}
+          <div className="space-y-2">
+            <Label htmlFor="email">{t("register.emailAddress")} *</Label>
+            <div className="relative">
+              <Mail className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
+              <Input
+                id="email"
+                type="email"
+                placeholder={t("register.emailPlaceholder")}
+                value={formData.email}
+                onChange={(e) => handleInputChange("email", e.target.value)}
+                className="pl-10 text-black"
+                required
+              />
+            </div>
+          </div>
+
+          {/* Phone */}
+          <div className="space-y-2">
+            <Label htmlFor="phone">{t("register.phoneNumber")} *</Label>
+            <div className="relative">
+              <Phone className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
+              <Input
+                id="phone"
+                type="tel"
+                placeholder={t("(+84)")}
+                value={formData.phone}
+                onChange={(e) => handleInputChange("phone", e.target.value)}
+                className="pl-10 text-black"
+                required
+              />
+            </div>
+          </div>
+
+          {/* Date of Birth */}
+          <div className="space-y-2">
+            <Label htmlFor="dateOfBirth">{t("register.dateOfBirth")} *</Label>
+            <div className="relative">
+              <Calendar className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
+              <Input
+                id="dateOfBirth"
+                type="date"
+                value={formData.dateOfBirth}
+                onChange={(e) => handleInputChange("dateOfBirth", e.target.value)}
+                className="pl-10 text-black"
+                required
+              />
+            </div>
+          </div>
+
+          {/* Password */}
+          <div className="space-y-2">
+            <Label htmlFor="password">{t("register.password")} *</Label>
+            <div className="relative">
+              <Lock className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
+              <Input
+                id="password"
+                type={showPassword ? "text" : "password"}
+                placeholder={t("register.passwordPlaceholder")}
+                value={formData.password}
+                onChange={(e) => handleInputChange("password", e.target.value)}
+                className="pl-10 pr-10 text-black"
+                required
+              />
+              <Button
+                type="button"
+                variant="ghost"
+                size="sm"
+                className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
+                onClick={() => setShowPassword(!showPassword)}
+              >
+                {showPassword ? <EyeOff className="h-4 w-4 text-muted-foreground" /> : <Eye className="h-4 w-4 text-muted-foreground" />}
+              </Button>
+            </div>
+          </div>
+
+          {/* Confirm Password */}
+          <div className="space-y-2">
+            <Label htmlFor="confirmPassword">{t("register.confirmPassword")} *</Label>
+            <div className="relative">
+              <Lock className="absolute left-3 top-1/2 transform -translate-y-1/2 text-muted-foreground h-5 w-5" />
+              <Input
+                id="confirmPassword"
+                type={showConfirmPassword ? "text" : "password"}
+                placeholder={t("register.confirmPasswordPlaceholder")}
+                value={formData.confirmPassword}
+                onChange={(e) => handleInputChange("confirmPassword", e.target.value)}
+                className="pl-10 pr-10 text-black"
+                required
+              />
+              <Button
+                type="button"
+                variant="ghost"
+                size="sm"
+                className="absolute right-0 top-0 h-full px-3 py-2 hover:bg-transparent"
+                onClick={() => setShowConfirmPassword(!showConfirmPassword)}
+              >
+                {showConfirmPassword ? <EyeOff className="h-4 w-4 text-muted-foreground" /> : <Eye className="h-4 w-4 text-muted-foreground" />}
+              </Button>
+            </div>
+          </div>
+
+          {/* Terms */}
+          <div className="flex items-center space-x-2">
+            <Checkbox
+              id="terms"
+              checked={agreeToTerms}
+              onCheckedChange={(checked) => setAgreeToTerms(checked as boolean)}
+            />
+            <label htmlFor="terms" className="text-sm leading-relaxed">
+              {t("register.agreeToTermsText")}{" "}
+              <Link to="/terms" className="text-primary hover:text-primary-dark">{t("register.termsOfService")}</Link>{" "}
+              {t("register.and")}{" "}
+              <Link to="/privacy" className="text-primary hover:text-primary-dark">{t("register.privacyPolicy")}</Link>
+            </label>
+          </div>
+
+          {/* Submit */}
+          <Button type="submit" className="w-full btn-hero" disabled={isLoading || !agreeToTerms}>
+            {isLoading ? t("common.creatingAccount") : t("common.createAccount")}
+          </Button>
+        </form>
+
+        <Separator />
+
+        {/* Sign in */}
+        <div className="text-center text-sm">
+          <span className="text-muted-foreground">{t("register.alreadyHaveAccount")} </span>
+          <Link to="/login" className="text-primary hover:text-primary-dark font-medium transition-colors">{t("register.signInHere")}</Link>
+        </div>
+      </CardContent>
+    </Card>
+  );
+};
