+++ frontend/src/pages/auth/Register/index.ts
@@ -1,0 +1,2 @@
+export { default } from "./Register";
+export * from "./types";
