+++ frontend/src/pages/auth/Register/types.ts
@@ -1,0 +1,11 @@
+export interface User {
+  id: string;
+  name: string;
+  email: string;
+  role: "customer" | "staff" | "admin";
+  stationId?: string;
+}
+
+export interface RegisterProps {
+  onRegister: (userData: User) => void;
+}
